+++ sdk/resourcemanager/containerservice/armcontainerservice/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Release History
 
+## 4.8.0 (2024-02-26)
+### Features Added
+
+- New struct `AgentPoolWindowsProfile`
+- New struct `ManagedClusterCostAnalysis`
+- New struct `ManagedClusterIngressProfile`
+- New struct `ManagedClusterIngressProfileWebAppRouting`
+- New struct `ManagedClusterMetricsProfile`
+- New field `IsDefault` in struct `KubernetesVersion`
+- New field `WindowsProfile` in struct `ManagedClusterAgentPoolProfile`
+- New field `WindowsProfile` in struct `ManagedClusterAgentPoolProfileProperties`
+- New field `IngressProfile`, `MetricsProfile` in struct `ManagedClusterProperties`
+
+
 ## 4.7.0 (2024-01-26)
 ### Features Added